migra endpoint deprecated para drfautoapi
