+++ sapl/api/urls.py
@@ -5,7 +5,7 @@ from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView, \
 from rest_framework.authtoken.views import obtain_auth_token
 
 from sapl.api.deprecated import SessaoPlenariaViewSet, \
-    AutoresProvaveisListView, AutoresPossiveisListView, AutorListView
+    AutoresProvaveisListView, AutorListView
 from sapl.api.views import AppVersionView, recria_token,\
     SaplApiViewSetConstrutor
 
@@ -39,8 +39,6 @@ urlpatterns_api_doc = [
 deprecated_urlpatterns_api = [
     url(r'^autor/provaveis',
         AutoresProvaveisListView.as_view(), name='autores_provaveis_list'),
-    url(r'^autor/possiveis',
-        AutoresPossiveisListView.as_view(), name='autores_possiveis_list'),
     url(r'^autor', AutorListView.as_view(), name='autor_list'),
 
 ]
