+++ sapl/api/views_compilacao.py
@@ -5,7 +5,7 @@ from drfautoapi.drfautoapi import ApiViewSetConstrutor, \
     customize, wrapper_queryset_response_for_drf_action
 
 
-CompilacaoApiViewSetConstrutor = ApiViewSetConstrutor.build_class(
+ApiViewSetConstrutor.build_class(
     [
         apps.get_app_config('compilacao')
     ]
