+++ sapl/protocoloadm/urls.py
@@ -124,12 +124,6 @@ urlpatterns_sistema = [
         include(TipoDocumentoAdministrativoCrud.get_urls())),
     url(r'^sistema/status-tramitacao-adm/',
         include(StatusTramitacaoAdministrativoCrud.get_urls())),
-
-    # FIXME: Usado para pesquisar autor- SOLUÇÃO-foi transformado em api/autor
-    # Melhor forma de fazer?
-    # Deve mudar de app?
-    # url(r'^protocoloadm/pesquisar-autor',
-    #    pesquisa_autores, name='pesquisar_autor'),
 ]
 
 urlpatterns = (urlpatterns_documento_administrativo +
