+++ sapl/templates/materia/autoria_form.html
@@ -5,15 +5,6 @@
 
 {% block extra_js %}
    <script language="Javascript">
-
-     function compare(a, b) {
-       if (a.text < b.text)
-         return -1;
-       if (a.text > b.text)
-         return 1;
-       return 0;
-     }
-
       $(document).ready(function() {
           $("#id_tipo_autor").change(function() {
               var tipo_selecionado = $("#id_tipo_autor").val();
@@ -22,19 +13,20 @@
               if (tipo_selecionado !== undefined && tipo_selecionado !== null) {
                 var json_data = {
                    tipo : tipo_selecionado,
-                   data_relativa : $("#id_data_relativa").val()
+                   data_relativa : $("#id_data_relativa").val(),
+                   o: 'nome',
+                   get_all: true
                 }
-                $.getJSON("/api/autor/possiveis", json_data, function(data){
+                $.getJSON("{% url 'sapl.api:autor-possiveis' %}", json_data, function(data){
                     if (data) {
-                      var results = data.sort(compare);
-                      if (results.length > 1) {
+                      if (data.length > 1) {
                         $("#id_autor").append("<option>-----</option>");
                       }
-                      $.each(results, function(idx, obj) {
+                      $.each(data, function(idx, obj) {
                           $("#id_autor")
                             .append($("<option></option>")
-                            .attr("value", obj.value)
-                            .text(obj.text));
+                            .attr("value", obj.id)
+                            .text(obj.nome));
                       });
                       $("#id_autor").val(autor_selecionado);
                     }
